// src/chessBoard/enums/colors.d.ts
export declare enum Colors {
    defaultBisque = "bisque",
    defaultBrown = "#4b2424",
    woodGreen = "green",
    woodBrown = "#8B4513",
    checkerRed = "red",
    checkerBlack = "black",
    skyBlue = "blue",
    skyWhite = "white",
    funKingdomPurple = "purple",
    funKingdomPink = "pink",
    classicBlack = "black",
    classicWhite = "white",
    vikingYellow = "yellow",
    vikingBlue = "blue",
    fireRed = "crimson",
    iceBlue = "lightblue"
}
export declare enum ColorSelect {
    selectId = "colorSelect",
    selectColor = "selectColor",
    default = "default",
    woodland = "woodland",
    checker = "checker",
    sky = "sky",
    funKingdom = "funKingdom",
    classic = "classic",
    viking = "viking",
    fireAndIce = "fireAndIce"
}